Refactor message handling and history management
- Standardized formatting and indentation across message handling functions in ContextManager and DatabaseHistoryProvider.
- Updated DatabaseHistoryProvider to utilize a DatabaseBackend for message storage and retrieval.
- Enhanced IConversationHistoryProvider interface with clear method definitions.
- Improved test coverage for DatabaseHistoryProvider and related components.
- Adjusted integration tests to ensure proper functionality with the new database structure.
