+++ src/core/brain/llm/messages/history/__test__/integration.test.ts
@@ -11,33 +11,33 @@ import type { InternalMessage } from '../../types.js';
 // NOTE: Integration test disabled due to module resolution issues with StorageManager/Logger in Vitest/ESM. To enable, uncomment imports above and run tests on compiled output or adjust test runner config.
 // NOTE: Test disabled due to unresolved StorageManager/createLogger in this environment.
 describe.skip('DatabaseHistoryProvider integration', () => {
-  let provider: DatabaseHistoryProvider;
-  let sessionId: string;
-  let message: InternalMessage;
+	let provider: DatabaseHistoryProvider;
+	let sessionId: string;
+	let message: InternalMessage;
 
-  const storageConfig = {
-  cache: { type: 'in-memory' },
-  database: { type: 'in-memory' }
-};
+	const storageConfig = {
+		cache: { type: 'in-memory' },
+		database: { type: 'in-memory' },
+	};
 
-beforeEach(() => {
-    sessionId = 'integration-session';
-    message = { role: 'user', content: 'integration test' };
-    provider = new DatabaseHistoryProvider(new StorageManager(storageConfig), createLogger());
-  });
+	beforeEach(() => {
+		sessionId = 'integration-session';
+		message = { role: 'user', content: 'integration test' };
+		provider = new DatabaseHistoryProvider(new StorageManager(storageConfig), createLogger());
+	});
 
-  it('should save and retrieve messages with real storage', async () => {
-    await provider.clearHistory(sessionId);
-    await provider.saveMessage(sessionId, message);
-    const history = await provider.getHistory(sessionId);
-    expect(history.length).toBeGreaterThan(0);
-    expect(history[history.length - 1]).toEqual(message);
-  });
+	it('should save and retrieve messages with real storage', async () => {
+		await provider.clearHistory(sessionId);
+		await provider.saveMessage(sessionId, message);
+		const history = await provider.getHistory(sessionId);
+		expect(history.length).toBeGreaterThan(0);
+		expect(history[history.length - 1]).toEqual(message);
+	});
 
-  it('should clear history with real storage', async () => {
-    await provider.saveMessage(sessionId, message);
-    await provider.clearHistory(sessionId);
-    const history = await provider.getHistory(sessionId);
-    expect(history.length).toBe(0);
-  });
+	it('should clear history with real storage', async () => {
+		await provider.saveMessage(sessionId, message);
+		await provider.clearHistory(sessionId);
+		const history = await provider.getHistory(sessionId);
+		expect(history.length).toBe(0);
+	});
 });
