+++ src/core/brain/llm/messages/history/factory.ts
@@ -1,8 +1,11 @@
-import { StorageManager } from '../../../storage/manager';
-import { Logger } from '../../../logger';
+import type { DatabaseBackend } from '../../../../storage/backend/database-backend.js';
+import { Logger } from '../../../../logger/index.js';
 import { DatabaseHistoryProvider } from './database.js';
 import { IConversationHistoryProvider } from './types.js';
 
-export function createDatabaseHistoryProvider(storage: StorageManager, logger: Logger): IConversationHistoryProvider {
-  return new DatabaseHistoryProvider(storage, logger);
+export function createDatabaseHistoryProvider(
+	database: DatabaseBackend,
+	logger: Logger
+): IConversationHistoryProvider {
+	return new DatabaseHistoryProvider(database, logger);
 }
